refactor: minor - move skill-\* files from 'back' to 'util'
