+++ util/skill-publish.js
@@ -3,12 +3,12 @@ const {
   TopicMessageSubmitTransaction,
 } = require('@hashgraph/sdk');
 
-const { client } = require('../util/sdk-client.js');
+const { client } = require('./sdk-client.js');
 const { skillVerify } = require('./skill-verify.js');
 const {
   addHash,
   serialise,
-} = require('../util/objects.js');
+} = require('./objects.js');
 
 async function skillPublish(topicId, accountId, userName, skillName) {
   if (typeof topicId === 'string') {
